backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
PyPDF2==3.0.1
pydantic==1.10.7
python-docx==0.8.11
python-dotenv==1.0.0
